+++ src/textual/_slug.py
@@ -34,10 +34,10 @@ NONLINGUAL: Final[str] = (
     r"\U000024C2-\U0001F251"
     r"\U00002702-\U000027B0"
     r"\U0001F1E0-\U0001F1FF"
-    r"\U0001F300-\U0001F5FF"
-    r"\U0001F600-\U0001F64F"
-    r"\U0001F680-\U0001F6FF"
-    r"\U0001f926-\U0001f937"
+    r"\U0001F300-\U0001F5FF"  # Miscellaneous Symbols And Pictographs
+    r"\U0001F600-\U0001F64F"  # Emoticons
+    r"\U0001F680-\U0001F6FF"  # Transport and Map Symbols
+    r"\U0001F900-\U0001F9FF"  # Supplemental Symbols and Pictographs
     r"\u200D"
     r"\u2640-\u2642"
 )
@@ -114,3 +114,27 @@ class TrackedSlugs:
         if used:
             slugged = f"{slugged}-{used}"
         return slugged
+
+
+VALID_ID_CHARACTERS = frozenset("abcdefghijklmnopqrstuvwxyz0123456789-")
+
+
+def slug_for_tcss_id(text: str) -> str:
+    """Produce a slug usable as a TCSS id from the given text.
+
+    Args:
+        text: Text.
+
+    Returns:
+        A slugified version of text suitable for use as a TCSS id.
+    """
+    is_valid = VALID_ID_CHARACTERS.__contains__
+    slug = "".join(
+        (character if is_valid(character) else "{:x}".format(ord(character)))
+        for character in text.casefold().replace(" ", "-")
+    )
+    if not slug:
+        return "_"
+    if slug[0].isdecimal():
+        return f"_{slug}"
+    return slug